some refactoring

+++ tests/douglas_pecker.cpp
@@ -14,6 +14,6 @@ TEST(douglas_peucker, simple1)
 	cout << v.len() << endl;
 	vector<point_2> chain( {point_2(0, 0), point_2(1, -1), point_2(2, -1), point_2(3, 0)} ), ans;
 	douglas_peucker(chain.begin(), chain.end(), 0.5, back_inserter(ans));
-	for(int i = 0; i < ans.size(); ++i)
+    for(size_t i = 0; i < ans.size(); ++i)
 		cout << ans[i] << endl;
 }
